+++ src/ebird/api/data/models/checklist.py
@@ -26,8 +26,9 @@ class Checklist(models.Model):
         verbose_name=_("published"),
     )
 
-    identifier = models.TextField(
-        unique=True,
+    identifier = models.CharField(
+        max_length=15,
+        primary_key=True,
         verbose_name=_("identifier"),
         help_text=_("The unique identifier for the checklist."),
     )