+++ src/ebird/api/data/models/observation.py
@@ -28,8 +28,9 @@ class Observation(models.Model):
         verbose_name=_("published"),
     )
 
-    identifier = models.TextField(
-        unique=True,
+    identifier = models.CharField(
+        max_length=15,
+        primary_key=True,
         verbose_name=_("identifier"),
         help_text=_("A global unique identifier for the observation."),
     )